remove cycle, fix test in CI
