+++ packages/datetime2/src/common/timezoneMetadata.ts
@@ -14,11 +14,9 @@
  * limitations under the License.
  */
 
-import { getTimezoneOffset } from "date-fns-tz";
-
 import { TIMEZONE_ITEMS } from "./timezoneItems";
 import { getTimezoneNames } from "./timezoneNameUtils";
-import { Timezone, TimezoneWithNames, TimezoneWithoutOffset } from "./timezoneTypes";
+import { Timezone, TimezoneWithNames } from "./timezoneTypes";
 
 /**
  * Given a timezone IANA code and an optional date object, retrieve additional metadata like its common name, offset,
@@ -30,28 +28,4 @@ export function getTimezoneMetadata(timezoneIanaCode: string, date?: Date): Time
         return undefined;
     }
     return getTimezoneNames(timezone, date);
-}
-
-/**
- * Augment hard-coded timezone information stored in this package with its current offset relative to UTC,
- * adjusted for daylight saving using the current date.
- *
- * @see https://github.com/marnusw/date-fns-tz#gettimezoneoffset
- */
-export function lookupTimezoneOffset(tz: TimezoneWithoutOffset, date?: Date): Timezone {
-    const offsetInMs = getTimezoneOffset(tz.ianaCode, date);
-    if (isNaN(offsetInMs)) {
-        throw new Error(`Unable to lookup offset for invalid timezone '${tz.ianaCode}'`);
-    }
-
-    const isPositiveOffset = offsetInMs >= 0;
-    const offsetInMinutes = Math.abs(offsetInMs) / 1000 / 60;
-    const offsetHours = Math.trunc(offsetInMinutes / 60)
-        .toString()
-        .padStart(2, "0");
-    const offsetMinutes = (offsetInMinutes % 60).toString().padEnd(2, "0");
-    return {
-        ...tz,
-        offset: `${isPositiveOffset ? "+" : "-"}${offsetHours}:${offsetMinutes}`,
-    };
 }
